[refactor] NewDisk PreviewList 컴포넌트 분리
