+++ src/utils/validations.ts
@@ -1,0 +1,5 @@
+// maxLength, minLenght, characters...
+
+export const DISK_NAME_MAX_LENGTH = 30;
+export const DISK_IMG_MAX_LENGTH = 4;
+export const DISK_CONTENT_MAX_LENGTH = 300;
